Parveidoju programmu vizuali skaistaku

+++ ai_Saldumi_11/Form1.cs
@@ -29,7 +29,7 @@ namespace ai_Saldumi_11
 
         private void Saldumi_aprekinat_button_Click(object sender, EventArgs e)
         {
-            double Naudas_daudzums_eur = Convert.ToInt32(this.Nauda_daudzums_textbox.Text);
+            double Naudas_daudzums_eur = Convert.ToDouble(this.Nauda_daudzums_textbox.Text);
             double Rezultats_eur = 0;
             Rezultats_eur = Naudas_daudzums_eur / 9.88;
             Rezultats_eur = Math.Round(Rezultats_eur, 2);
